feat(checkout): CHECKOUT-10104 refactor interfaces
